Wave Manager + Equipment
- Added a wave manager and passed variables from GameManager to it
- Started adding equipement for the character

+++ Assets/Scripts/Units/Hero/Archetypes/Archer.cs
@@ -6,7 +6,7 @@ public class Archer : Hero
     [Header("Effect")]
     [SerializeField] private GameObject arrow;
 
-    protected override IEnumerator AttackCR(ushort damage)
+    protected override IEnumerator AttackCR(int damage)
     {
         rb.velocity = Vector2.zero;
         anim.Play("Attack");
