+++ Assets/Scripts/Units/Hero/Archetypes/Assassin.cs
@@ -9,7 +9,7 @@ public class Assassin : Hero
 
 
 
-    protected override IEnumerator AttackCR(ushort damage)
+    protected override IEnumerator AttackCR(int damage)
     {
         rb.velocity = Vector2.zero;
         anim.Play("Attack");
