+++ Assets/Scripts/Units/Hero/Archetypes/Warrior.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 public class Warrior : Hero
 {
-    protected override IEnumerator AttackCR(ushort damage)
+    protected override IEnumerator AttackCR(int damage)
     {
         rb.velocity = Vector2.zero;
         anim.Play("Attack");
